Update project.sql
Considering if we should put make, model and two integers to represent an upper and lower bound on the years of compatible cars. seems like an easier to implement solution without losing detail. What do you think? Did not do it in this update, just started the INSERT INTO statements. To make this a little easier we might want to make the assumption that we work with commonly damaged parts among the most popular vehicles in areas with frequent collisions (Like top 10 - 12)